pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "policyengine_uk_data"
version = "1.36.2"
description = "A package to create representative microdata for the UK."
readme = "README.md"
authors = [
    {name = "PolicyEngine", email = "hello@policyengine.org"},
]
license = {file = "LICENSE"}
requires-python = ">=3.13"
dependencies = [
    "policyengine-core>=3.19.4",
    "requests",
    "httpx>=0.27.0",
    "tqdm",
    "tabulate",
    "huggingface_hub",
    "policyengine",
    "google-cloud-storage",
    "google-auth",
    "policyengine-uk>=2.43.5",
    "microcalibrate>=0.18.0",
    "microimpute>=1.0.1",
    "black>=25.1.0",
    "rich>=13.0.0",
    "odfpy",
    "pandas",
    "openpyxl",
    "dagster>=1.9.0",
    "dagster-webserver>=1.9.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "torch",
    "tables",
    "furo",
    "yaml-changelog>=0.1.7",
    "itables",
    "quantile-forest",
    "build",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["policyengine_uk_data*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"policyengine_uk_data" = ["**/*"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["policyengine_uk_data/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 79
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.dagster]
module_name = "policyengine_uk_data.definitions" 
